requirements.txt
Django==5.2.5
Pillow==11.3.0